+++ dftimewolf/lib/containers/__init__.py
@@ -1,3 +1,3 @@
-"""Make datatypes available here."""
+"""Make containers available here."""
 from .report import Report
 from .threat_intelligence import ThreatIntelligence
