+++ tests/lib/containers/threat_intelligence.py
@@ -1,20 +1,22 @@
 # -*- coding: utf-8 -*-
-"""Tests for the Report attribute containers."""
+"""Tests for the ThreatIntelligence attribute containers."""
 
 from __future__ import unicode_literals
 
 import unittest
 
-from dftimewolf.lib import datatypes
+from dftimewolf.lib import containers
 
 class ReportDataTest(unittest.TestCase):
   """Tests for the Report data attribute container."""
 
   def testGetAttributeNames(self):
     """Tests the GetAttributeNames function."""
-    attribute_container = datatypes.Report(module_name='name', text='text')
+    attribute_container = containers.ThreatIntelligence(
+        name='name',
+        indicator='.*')
 
-    expected_attribute_names = ['module_name', 'text']
+    expected_attribute_names = ['indicator', 'name']
 
     attribute_names = sorted(attribute_container.GetAttributeNames())
